+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Implemented reverse WebFinger for portable actors.
+- Added account limits to Mastodon API `Instance` entity.
 
 ### Changed